// src/contracts/api.ts
export interface paths {
    "/api/personas": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /** ペルソナの作成 */
        post: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody: {
                content: {
                    "application/json": components["schemas"]["PersonaCreate"];
                };
            };
            responses: {
                /** @description 作成成功 */
                201: {
                    headers: {
                        Location: string;
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            id: string;
                        };
                    };
                };
                /** @description 不正なリクエスト */
                400: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        /**
                         * @example {
                         *       "message": "入力内容が正しくありません"
                         *     }
                         */
                        "application/json": unknown;
                    };
                };
                /** @description An unexpected error occurred */
                default: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/personas/{personaId}/projects": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * ペルソナのプロジェクトの一覧
         * @description ペルソナのプロジェクトの一覧。プロジェクトは試験日の昇順。
         */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    personaId: components["parameters"]["PersonaId"];
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description プロジェクトの一覧 */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["Project"][];
                    };
                };
// ...
            };
        };
// ...
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        PersonaCreate: {
            /** @example 2 */
            weekdayHours: number;
            /** @example 5 */
            weekendHours: number;
            /**
             * @example インプット先行パターン
             * @enum {string}
             */
            learningPattern: "インプット先行パターン" | "アウトプット先行パターン";
        };
        Project: {
            /** Format: uuid */
            projectId: string;
            /** Format: uuid */
            personaId: string;
            certificationName: string;
            /** Format: date */
            startDate: string;
            /** Format: date */
            examDate: string;
            /** @enum {string} */
            baseMaterial: "TEXTBOOK" | "VIDEO";
        };
        ProjectCreate: {
            /** @example AWS Solutions Architect Associate */
            certificationName: string;
            /**
             * Format: date
             * @example 2025-10-01
             */
            startDate: string;
            /**
             * Format: date
             * @example 2025-12-15
             */
            examDate: string;
            /**
             * @example TEXTBOOK
             * @enum {string}
             */
            baseMaterial: "TEXTBOOK" | "VIDEO";
        };
        Step: {
            /** Format: uuid */
            stepId?: string;
            /** Format: uuid */
            projectId?: string;
            title?: string;
            theme?: string;
            /** Format: date */
            startDate?: string;
            /** Format: date */
            endDate?: string;
            /** @description 小さい値から昇順に並ぶシリアル番号 */
            index?: number;
        };
        StepCreate: {
            title: string;
            theme: string;
            /** @description 小さい値から昇順に並ぶシリアル番号 */
            index: number;
        };
        Task: {
            /** Format: uuid */
            taskId?: string;
            /** Format: uuid */
            stepId?: string;
            title?: string;
            description?: string;
            /** Format: date */
            startDate?: string;
            /** Format: date */
            endDate?: string;
            /** @enum {string} */
            taskStatus?: "undo" | "doing" | "done" | "blocked";
        };
        Weight: {
            /** Format: uuid */
            weightId?: string;
            /** Format: uuid */
            projectId?: string;
            area?: string;
            weightPercent?: number;
        };
        WeightCreate: {
            area: string;
            /** @example 30 */
            weightPercent: number;
        };
    };
    responses: never;
    parameters: {
        PersonaId: string;
        ProjectId: string;
        StepId: string;
        TaskId: string;
        ScopeId: string;
        WeightId: string;
    };
    requestBodies: never;
    headers: never;
    pathItems: never;
}